requirements.txt
numpy
scipy
soundfile
psycopg2-binary
